Rename gentx-5880b4ac41a4ebad9ce734919ce8452e5f8f01e0.json to gentx-59.json
I submitted wrong census pubkey before the valid one and it produced wrong GenTX filename. 
If current filename will not going to have any issue in the future, that's alright but I recommend to to homogenize with others. 
The content of my GenTX remains the same. I'm sorry for not reading the instruction correctly.